Export from index.ts, small update to Divide operator
